Moved api related hooks into hooks/api/
