+++ jiu-jitsu-notes/src/hooks/api/usePositionsApi.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import positionsApi from "../api/positions";
+import positionsApi from "../../api/positions";
 
 const usePositionsApi = () => {
   const [positions, setPositions] = useState([]);
